feat(QuestionDetail): Recorder 컴포넌트 분리
